+++ examples/default/src/App.tsx
@@ -1,18 +1,20 @@
-import React, { useEffect } from 'react';
-import { StyleSheet } from 'react-native';
+import React, { useEffect, useState } from 'react';
+import { ActivityIndicator, StyleSheet } from 'react-native';
 
 import { GestureHandlerRootView } from 'react-native-gesture-handler';
 import { NavigationContainer, useNavigationContainerRef } from '@react-navigation/native';
+import type { SessionMetadata } from 'instabug-reactnative';
 import Instabug, {
   CrashReporting,
   InvocationEvent,
+  LaunchType,
   LogLevel,
+  NetworkInterceptionMode,
+  NetworkLogger,
   ReproStepsMode,
   SessionReplay,
-  LaunchType,
   OverAirUpdateServices,
 } from 'instabug-reactnative';
-import type { SessionMetadata } from 'instabug-reactnative';
 import { NativeBaseProvider } from 'native-base';
 
 import { RootTabNavigator } from './navigation/RootTab';
@@ -39,33 +41,56 @@ export const App: React.FC = () => {
 
   const navigationRef = useNavigationContainerRef();
 
+  const [isInstabugInitialized, setIsInstabugInitialized] = useState(false);
+
+  const initializeInstabug = async () => {
+    try {
+      SessionReplay.setSyncCallback((data) => shouldSyncSession(data));
+
+      await Instabug.init({
+        token: 'deb1910a7342814af4e4c9210c786f35',
+        invocationEvents: [InvocationEvent.floatingButton],
+        debugLogsLevel: LogLevel.verbose,
+        networkInterceptionMode: NetworkInterceptionMode.native,
+        overAirVersion: { service: OverAirUpdateServices.codePush, version: '1.0.0' },
+      });
+
+      CrashReporting.setNDKCrashesEnabled(true);
+      Instabug.setReproStepsConfig({ all: ReproStepsMode.enabled });
+
+      setIsInstabugInitialized(true); // Set to true after initialization
+    } catch (error) {
+      console.error('Instabug initialization failed:', error);
+      setIsInstabugInitialized(true); // Proceed even if initialization fails
+    }
+  };
+
   useEffect(() => {
-    SessionReplay.setSyncCallback((data) => shouldSyncSession(data));
-
-    Instabug.init({
-      token: 'deb1910a7342814af4e4c9210c786f35',
-      invocationEvents: [InvocationEvent.floatingButton],
-      debugLogsLevel: LogLevel.verbose,
-      overAirVersion: { service: OverAirUpdateServices.codePush, version: '1.0.0' },
+    initializeInstabug().then(() => {
+      NetworkLogger.setNetworkDataObfuscationHandler(async (networkData) => {
+        networkData.url = `${networkData.url}/JS/Obfuscated`;
+        return networkData;
+      });
+      // NetworkLogger.setRequestFilterExpression('false');
     });
-    CrashReporting.setNDKCrashesEnabled(true);
-
-    Instabug.setReproStepsConfig({
-      all: ReproStepsMode.enabled,
-    });
-  }, []);
+  });
 
   useEffect(() => {
+    // @ts-ignore
     const unregisterListener = Instabug.setNavigationListener(navigationRef);
 
     return unregisterListener;
   }, [navigationRef]);
 
+  if (!isInstabugInitialized) {
+    return <ActivityIndicator size="large" color="#0000ff" style={styles.loading} />;
+  }
+
   return (
     <GestureHandlerRootView style={styles.root}>
       <NativeBaseProvider theme={nativeBaseTheme}>
         <QueryClientProvider client={queryClient}>
-          <NavigationContainer theme={navigationTheme} ref={navigationRef}>
+          <NavigationContainer onStateChange={Instabug.onStateChange} theme={navigationTheme}>
             <RootTabNavigator />
           </NavigationContainer>
         </QueryClientProvider>
@@ -77,5 +102,10 @@ export const App: React.FC = () => {
 const styles = StyleSheet.create({
   root: {
     flex: 1,
+  },
+  loading: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
   },
 });