+++ src/native/NativeInstabug.ts
@@ -157,6 +157,8 @@ export interface InstabugNativeModule extends NativeModule {
 
   // W3C Feature Flags Listener for Android
   registerW3CFlagsChangeListener(): void;
+
+  setOnFeaturesUpdatedListener(handler?: (params: any) => void): void; // android only
   enableAutoMasking(autoMaskingTypes: AutoMaskingType[]): void;
 }
 
@@ -164,6 +166,7 @@ export const NativeInstabug = NativeModules.Instabug;
 
 export enum NativeEvents {
   PRESENDING_HANDLER = 'IBGpreSendingHandler',
+  IBG_ON_FEATURES_UPDATED_CALLBACK = 'IBGOnFeatureUpdatedCallback',
   ON_W3C_FLAGS_CHANGE = 'IBGOnNewW3CFlagsUpdateReceivedCallback',
 }