+++ test/mocks/mockInstabug.ts
@@ -75,6 +75,7 @@ const mockInstabug: InstabugNativeModule = {
   isW3CaughtHeaderEnabled: jest.fn(),
   registerW3CFlagsChangeListener: jest.fn(),
   setNetworkLogBodyEnabled: jest.fn(),
+  setOnFeaturesUpdatedListener: jest.fn(),
   enableAutoMasking: jest.fn(),
 };